+++ changelog.md
@@ -18,6 +18,7 @@ Enhancements
 8. Moved Audit information to top of record, 
    collapsed by default.
 
+
 Bug Fixes
 ---------
 1. Substructure searches sometimes attempted to render
@@ -42,6 +43,9 @@ Bug Fixes
 7. Don't fail text indexing on empty strings. We now
    allow empty strings to be ignored from indexing
    or to add them with a special "EMTPY" keyword.
+8. Removed search from protein thumb if no subunits
+9. Added validation rule to proteins, not to allow
+   records with no subunits
 
 
 Deeper look